+++ opencv-extra/include/aruco-namespace.hpp
@@ -1,0 +1,6 @@
+#ifndef __OPENCV_ARUCO_NAMESPACE_H__
+#define __OPENCV_ARUCO_NAMESPACE_H__
+
+using namespace cv::aruco;
+
+#endif /* __OPENCV_ARUCO_NAMESPACE_H__ */